Triggerlänge pro Spur

+++ include/core/app_config.h
@@ -10,7 +10,10 @@ struct Bank {
   // Mit einem festen Array lassen sich die Namen sicher mit strncpy_P() auslesen.
   const char name[19];
   uint8_t stepsPerBar;            // Schritte pro Takt: 16 für 4/4, 12 für 3/4
-  uint16_t patterns[9][4];  // Index 0..7 = Kanaele, Index 8 = Open Hi-Hat (gleicher Pin wie Kanal 6)
+  uint32_t patterns[9][4];  // 0..7 = Kanaele, 8 = Open Hi-Hat
+                            // Jedes Pattern ist 32 Bit:
+                            // - Bits 0-15 = Hit-Maske (X oder O = 1)
+                            // - Bits 16-31 = Dead-Node-Maske (nur O = 1)
 };
 
 static const int bankAddr = 0;
